Configuration variables now loaded from settings.py
Optional environment variable GROWLPROXY_SETTINGS can also be used to
specifiy the path to a .py file to load settings from.

The model code was updated to load the database from the file specified
by the DATABASE setting.  Currently only sqlite3 databases are
supported, but that could fairly easily be updated.