Fixed configuration of database filename
Previously the database engine was always being set when db.py was
imported, so the database name was pretty much always the default.

This changeset updates it so that the values are initialised by a call
to a new function InitDb.  This function will also optionally create the
database structure.  Have also update runui.py and the base ui test case
to use InitDb